Conflicts:
	addons/ofxAndroid/src/ofAppAndroidWindow.cpp

+++ scripts/dev/create_package.sh
@@ -1,9 +1,16 @@
 #!/bin/bash
 # $1 -> platform: win_cb, linux, linux64, vs, osx, ios, all
 # $2 -> version number: 006
+#
+# This script removes folders clones the openFrameworks repo 
+# and deletes parts of it to create the final package.
+# Do not try to modify it to run over your local install of
+# openFrameworks or it'll remove it along with any projects it
+# might contain.
 
 platform=$1
 version=$2
+of_root=$(readlink -f "$(dirname "$(readlink -f "$0")")/../..")
 
 if [ $# -eq 3 ]; then
 branch=$3
@@ -11,8 +18,8 @@ else
 branch=stable
 fi
 
-REPO=https://github.com/openframeworks/openFrameworks
-REPO_ALIAS=upstreamhttps
+REPO=../..
+REPO_ALIAS=originlocal
 BRANCH=$branch
 
 PG_REPO=https://github.com/openframeworks/projectGenerator.git
@@ -21,7 +28,7 @@ PG_BRANCH=master
 
 hostArch=`uname`
 
-if [ "$platform" != "win_cb" ] && [ "$platform" != "linux" ] && [ "$platform" != "linux64" ] && [ "$platform" != "linuxarmv6l" ] && [ "$platform" != "linuxarmv7l" ] && [ "$platform" != "vs" ] && [ "$platform" != "osx" ] && [ "$platform" != "android" ] && [ "$platform" != "ios" ] && [ "$platform" != "all" ]; then
+if [ "$platform" != "win_cb" ] && [ "$platform" != "linux" ] && [ "$platform" != "linux64" ] && [ "$platform" != "linuxarmv6l" ] && [ "$platform" != "linuxarmv7l" ] && [ "$platform" != "vs" ] && [ "$platform" != "osx" ] && [ "$platform" != "android" ] && [ "$platform" != "ios" ]; then
     echo usage: 
     echo ./create_package.sh platform version
     echo platform:
@@ -40,6 +47,13 @@ if [ "$version" == "" ]; then
     exit 1
 fi
 
+echo
+echo
+echo
+echo --------------------------------------------------------------------------
+echo "Creating package $version for $platform"
+echo --------------------------------------------------------------------------
+echo
 
 libsnotinmac="glu quicktime videoInput kiss"
 libsnotinlinux="quicktime videoInput glut glu cairo glew openssl rtAudio"
@@ -48,20 +62,34 @@ libsnotinmingw="kiss glut cairo glew openssl"
 libsnotinandroid="glut quicktime videoInput fmodex glee rtAudio kiss cairo"
 libsnotinios="glut quicktime videoInput fmodex glee rtAudio kiss cairo"
 
-rm -rf openFrameworks
-git clone $REPO --depth=1 --branch=$BRANCH
+# This script removes folders clones the openFrameworks repo 
+# and deletes parts of it to create the final package.
+# Do not try to modify it to run over your local install of
+# openFrameworks or it'll remove it along with any projects it
+# might contain.
+# Instead we download it in a folder with a different name so it's
+# safe to delete it completely at the end
+pkgfolder=openFrameworks_pkg_creation
+
+rm -rf ${pkgfolder}
+echo "Cloning OF from $REPO $BRANCH" 
+git clone $REPO --depth=1 --branch=$BRANCH ${pkgfolder} 2> /dev/null
 gitfinishedok=$?
 if [ $gitfinishedok -ne 0 ]; then
     echo "Error connecting to github"
     exit
 fi
 
+cd ${pkgfolder}
+packageroot=$PWD
 
-if [ "$platform" != "linux" ] || [ "$platform" != "linux64" ] || [ "$platform" != "linuxarmv6l" ] || [ "$platform" != "linuxarmv7l" ]; then
-    cd openFrameworks
-    packageroot=$PWD
-    cd apps
-    git clone $PG_REPO --depth=1 --branch=$PG_BRANCH
+cd apps
+echo "Cloning project generator from $PG_REPO $PG_BRANCH" 
+git clone $PG_REPO --depth=1 --branch=$PG_BRANCH 2> /dev/null
+gitfinishedok=$?
+if [ $gitfinishedok -ne 0 ]; then
+    echo "Error connecting to github"
+    exit
 fi
 
 cd $packageroot
@@ -70,7 +98,6 @@ cd $packageroot
 function deleteCodeblocks {
     #delete codeblock files
     rm *.cbp
-    rm *.sh
     rm *.workspace
 }
 
@@ -102,46 +129,43 @@ function deleteEclipse {
 
 
 function createProjectFiles {
-    if [ "$pkg_platform" == "win_cb" ]; then
-	    # copy all examples to pkg_ofroot
-	    cp -Rf $packageroot/examples $pkg_ofroot
-	
+    if [ "$pkg_platform" == "win_cb" ]; then	
+        echo "Creating project files"
 	    # copy config.make and Makefile into every subfolder
-	    for d in `find ${pkg_ofroot}/examples -maxdepth 2 -type d`; do
-	      cp $packageroot/scripts/win_cb/template/config.make ${d}
-	      cp $packageroot/scripts/win_cb/template/Makefile ${d}
+	    for example in $pkg_ofroot/examples/*/*; do
+	        cp $pkg_ofroot/scripts/templates/win_cb/config.make ${example}
+	        cp $pkg_ofroot/scripts/templates/win_cb/Makefile ${example}
 	    done
-	
-	    # remove config.make and Makefile from level 1
-        for d in `find ${pkg_ofroot}/examples -maxdepth 1 -type d`; do
-	      rm "${d}/config.make"
-	      rm "${d}/Makefile"
-	    done
-	
-	    # remove config.make and Makefile from level 0
-	    rm "$pkg_ofroot/examples/config.make"
-	    rm "$pkg_ofroot/examples/Makefile"
-    else
-        PG_OF_PATH=$pkg_ofroot projectGenerator --recursive -p ${pkg_platform} $pkg_ofroot/examples
+    elif [ "$pkg_platform" != "android" ]; then
+        cd ${main_ofroot}/apps/projectGenerator
+        git pull origin master
+        cd commandLine
+        echo "Recompiling command line PG"
+        PROJECT_OPTIMIZATION_CFLAGS_RELEASE=-O3 make -j2 > /dev/null
+        cd ${pkg_ofroot}
+        echo "Creating project files for $pkg_platform"
+        ${main_ofroot}/apps/projectGenerator/commandLine/bin/projectGenerator --recursive -p ${pkg_platform} -o $pkg_ofroot $pkg_ofroot/examples > /dev/null
     fi
+    cd ${pkg_ofroot}
 }
 
 function createPackage {
     pkg_platform=$1
     pkg_version=$2
     pkg_ofroot=$3
+    main_ofroot=$4
     
     #remove previously created package 
     cd $pkg_ofroot/..
 	rm -Rf of_v${pkg_version}_${pkg_platform}.*
 	rm -Rf of_v${pkg_version}_${pkg_platform}_*
-    echo "creating package $pkg_platform $version in $pkg_ofroot"
+    echo "Creating package $pkg_platform $version in $pkg_ofroot"
     
     #remove devApps folder
     rm -r $pkg_ofroot/apps/devApps
     
     #remove projectGenerator folder
-    if [ "$pkg_platform" != "linux" ] && [ "$pkg_platform" != "linux64" ]; then
+    if [ "$pkg_platform" = "android" ] || [ "$pkg_platform" = "win_cb" ]; then
     	rm -rf $pkg_ofroot/apps/projectGenerator
     fi
 
@@ -201,7 +225,12 @@ function createPackage {
         rm -Rf gl/gpuParticleSystemExample
         rm -Rf gl/vboMeshDrawInstancedExample
         rm -Rf gl/shaderExample
-        
+        rm -Rf gl/computeShaderParticlesExample
+        rm -Rf gl/computeShaderTextureExample
+        rm -Rf gl/pixelBufferExample
+        rm -Rf gl/textureBufferInstancedExample
+        rm -Rf gl/threadedPixelBufferExample
+  
         rm -Rf utils/systemSpeakExample
         rm -Rf utils/fileBufferLoadingCSVExample
         
@@ -221,6 +250,8 @@ function createPackage {
 	
 	if [ "$pkg_platform" == "osx" ]; then
 	    rm -Rf gles
+	    rm -Rf gl/computeShaderParticlesExample
+	    rm -Rf gl/computeShaderTextureExample
 	fi
 	
 	
@@ -273,55 +304,86 @@ function createPackage {
     
     
 	#download and uncompress PG
-	cd $pkg_ofroot
-	rm -rf projectGenerator
-    #if [ "$pkg_platform" = "win_cb" ]; then
-	#	rm projectGenerator_wincb.zip
-	#	wget http://www.openframeworks.cc/pgSimple/projectGenerator_wincb.zip
-	#	unzip projectGenerator_wincb.zip
-	#	rm projectGenerator_wincb.zip
-	#	rm -Rf __MACOSX
-	#fi
+	echo "Creating projectGenerator"
+	mkdir -p $HOME/.tmp
+	export TMPDIR=$HOME/.tmp
     if [ "$pkg_platform" = "vs" ]; then
-		rm projectGenerator_vs.zip
-		wget http://192.237.185.151/projectGenerator/projectGenerator_vs.zip
-		unzip projectGenerator_vs.zip
-		rm projectGenerator_vs.zip
-		rm -Rf __MACOSX
+		cd ${pkg_ofroot}/apps/projectGenerator/projectGeneratorElectron
+		npm install > /dev/null
+		npm run build:vs > /dev/null
+		mv dist/projectGenerator-win32-ia32 ${pkg_ofroot}/projectGenerator-vs
+		cd ${pkg_ofroot}
+		rm -rf apps/projectGenerator
+		cd ${pkg_ofroot}/projectGenerator-vs/resources/app/app/
+		wget http://192.237.185.151/projectGenerator/projectGenerator-vs.zip 2> /dev/null
+		unzip projectGenerator-vs.zip 2> /dev/null
+		rm projectGenerator-vs.zip
+		cd ${pkg_ofroot}
+		sed -i "s/osx/vs/g" projectGenerator-vs/resources/app/settings.json
 	fi
     if [ "$pkg_platform" = "osx" ]; then
-		rm projectGenerator_osx.zip
-		wget http://192.237.185.151/projectGenerator/projectGenerator_osx.zip
-		unzip projectGenerator_osx.zip
-		rm projectGenerator_osx.zip
-		rm -Rf __MACOSX
+		cd ${pkg_ofroot}/apps/projectGenerator/projectGeneratorElectron
+		npm install > /dev/null
+		npm run build:osx > /dev/null
+		mv dist/projectGenerator-darwin-x64 ${pkg_ofroot}/projectGenerator-osx
+		cd ${pkg_ofroot}
+		rm -rf apps/projectGenerator
+		wget http://192.237.185.151/projectGenerator/projectGenerator_osx -O projectGenerator-osx/projectGenerator.app/Contents/Resources/app/app/projectGenerator 2> /dev/null
+		sed -i "s/osx/osx/g" projectGenerator-osx/projectGenerator.app/Contents/Resources/app/settings.json
 	fi
     if [ "$pkg_platform" = "ios" ]; then
-		rm projectGenerator_ios.zip
-		wget http://192.237.185.151/projectGenerator/projectGenerator_ios.zip
-		unzip projectGenerator_ios.zip
-		rm projectGenerator_ios.zip
-		rm -Rf __MACOSX
+		cd ${pkg_ofroot}/apps/projectGenerator/projectGeneratorElectron
+		npm install > /dev/null
+		npm run build:osx > /dev/null
+		mv dist/projectGenerator-darwin-x64 ${pkg_ofroot}/projectGenerator-ios
+		cd ${pkg_ofroot}
+		rm -rf apps/projectGenerator
+		wget http://192.237.185.151/projectGenerator/projectGenerator_osx -O projectGenerator-ios/projectGenerator.app/Contents/Resources/app/app/projectGenerator 2> /dev/null
+		sed -i "s/osx/ios/g" projectGenerator-ios/projectGenerator.app/Contents/Resources/app/settings.json
+	fi
+	
+	if [ "$pkg_platform" = "linux" ]; then
+		cd ${pkg_ofroot}/apps/projectGenerator/projectGeneratorElectron
+		npm install > /dev/null
+		npm run build:linux > /dev/null
+		mv dist/projectGenerator-linux-ia32 ${pkg_ofroot}/projectGenerator-linux
+		cd ${pkg_ofroot}
+		sed -i "s/osx/linux/g" projectGenerator-linux/resources/app/settings.json
+	fi
+	
+	if [ "$pkg_platform" = "linux64" ]; then
+		cd ${pkg_ofroot}/apps/projectGenerator/projectGeneratorElectron
+		npm install > /dev/null
+		npm run build:linux64 > /dev/null
+		mv dist/projectGenerator-linux-x64 ${pkg_ofroot}/projectGenerator-linux64
+		cd ${pkg_ofroot}
+		sed -i "s/osx/linux64/g" projectGenerator-linux64/resources/app/settings.json
 	fi
 	
 	# linux remove other platform projects from PG source and copy ofxGui
 	if [ "$pkg_platform" = "linux" ] || [ "$pkg_platform" = "linux64" ] || [ "$pkg_platform" = "linuxarmv6l" ] || [ "$pkg_platform" = "linuxarmv7l" ]; then
-		cd apps/projectGenerator/projectGeneratorSimple
+	    cd ${pkg_ofroot}
+		mv apps/projectGenerator/commandLine .
+		rm -rf apps/projectGenerator
+		mkdir apps/projectGenerator
+		mv commandLine apps/projectGenerator/
+		cd apps/projectGenerator/commandLine
 		deleteCodeblocks
 		deleteVS
 		deleteXcode
-		rm -Rf .git*
 	fi
 
     #delete libraries for other platforms
+    echo "Deleting core libraries from other platforms"
     cd $pkg_ofroot/libs  
     for lib in $( find . -maxdepth 1 -mindepth 1 -type d )
     do
-        echo $PWD
-        echo deleting $lib/lib
-        cd $lib/lib
-        rm -Rf $otherplatforms
-        cd $pkg_ofroot/libs 
+        if [ -d $lib/lib ]; then
+            #echo deleting $lib/lib
+            cd $lib/lib
+            rm -Rf $lib/lib/$otherplatforms
+            cd $pkg_ofroot/libs
+        fi
     done
     if [ "$pkg_platform" = "osx" ]; then
         rm -Rf $libsnotinmac
@@ -330,7 +392,7 @@ function createPackage {
     elif [ "$pkg_platform" = "win_cb" ]; then
         rm -Rf $libsnotinmingw
     elif [ "$pkg_platform" = "vs" ]; then
-        rm -Rf $libsnotinwindows
+        rm -Rf $libsnotinvs
     elif [ "$pkg_platform" = "android" ]; then
         rm -Rf $libsnotinandroid
     elif [ "$pkg_platform" = "ios" ]; then
@@ -338,11 +400,11 @@ function createPackage {
     fi
     
     cd ${pkg_ofroot}/addons
+    echo "Deleting addon libraries from other platforms"
     for lib in $( ls -d */libs/*/lib/ )
     do
         cd ${lib}
-        echo $PWD
-        echo deleting $lib
+        #echo deleting $lib
         rm -Rf $otherplatforms
         cd $pkg_ofroot/addons
     done
@@ -373,13 +435,15 @@ function createPackage {
 	fi
 	
 	#delete unit tests by now
-	rm -Rf ${pkg_root}/tests
+	rm -Rf ${pkg_ofroot}/tests
 
 	#delete eclipse projects
 	if [ "$pkg_platform" != "android" ] && [ "$pkg_platform" != "linux" ] && [ "$pkg_platform" != "linux64" ] && [ "$pkg_platform" != "linuxarmv6l" ] && [ "$pkg_platform" != "linuxarmv7l" ]; then
 		cd ${pkg_ofroot}
 		deleteEclipse
-		rm -R libs/openFrameworks/.settings
+		if [ -f libs/openFrameworks/.settings ]; then
+    		rm -R libs/openFrameworks/.settings
+    	fi
 	fi
 	
 	#android, move paths.default.make to paths.make
@@ -403,12 +467,12 @@ function createPackage {
 
 	cd ${pkg_ofroot}/libs
 	#delete specific include folders non-android
-	if [ "$pkg_platform" != "android" ]; then
+	if [ "$pkg_platform" != "android" ] && [ -d */include_android ]; then
 		rm -Rf $( ls -d */include_android )
 	fi
 
 	#delete specific include folders for non-ios
-	if [ "$pkg_platform" != "ios" ]; then
+	if [ "$pkg_platform" != "ios" ] && [ -d */include_ios ]; then
 		rm -Rf $( ls -d */include_ios )
 	fi
 
@@ -420,10 +484,6 @@ function createPackage {
     #delete dynamic libraries for other platforms
     cd $pkg_ofroot/export
     rm -Rf $otherplatforms
-    if [ "$pkg_platform" = "osx" ]; then
-        cd $pkg_ofroot
-        rmdir export
-    fi
 
     #delete scripts
     cd $pkg_ofroot/scripts
@@ -436,11 +496,6 @@ function createPackage {
     #delete omap4 scripts for non armv7l
 	if [ "$pkg_platform" = "linux64" ] || [ "$pkg_platform" = "linux" ] || [ "$pkg_platform" = "linuxarmv6l" ]; then
 	    rm -Rf linux/ubuntu-omap4
-	fi
-	
-    #delete armv6 scripts for non armv6l
-	if [ "$pkg_platform" = "linux64" ] || [ "$pkg_platform" = "linux" ] || [ "$pkg_platform" = "linuxarmv7l" ]; then
-	    rm -Rf linux/debian_armv6l
 	fi
 	
 	if [ "$pkg_platform" == "ios" ]; then
@@ -464,23 +519,9 @@ function createPackage {
 	if [ "$pkg_platform" != "osx" ] && [ "$pkg_platform" != "ios" ]; then
 		rm -Rf "xcode templates"
 	fi
-	
-	
-	#download and copy OF compiled
-	#cd $pkg_ofroot/libs/openFrameworksCompiled/lib/${pkg_platform}
-    	#if [ "$pkg_platform" = "win_cb" ]; then
-	#	wget http://openframeworks.cc/git_pkgs/OF_compiled/${pkg_platform}/openFrameworks.lib
-	#	wget http://openframeworks.cc/git_pkgs/OF_compiled/${pkg_platform}/openFrameworksDebug.lib
-	#fi
-
-
-    #if snow leopard change 10.4u to 10.5
-    #if [ $runOSXSLScript = 1 ]; then
-    #    cd $pkg_ofroot
-    #    echo "replacing 10.4u with 10.5 for snow leopard"
-    #    find . -name '*.pbxproj' | xargs perl -pi -e 's/10\.4u/10\.5/g'
-    #    pkg_platform="osxSL"
-    #fi
+    echo ----------------------------------------------------------------------
+    echo
+    echo
     
     #choose readme
     cd $pkg_ofroot
@@ -501,64 +542,66 @@ function createPackage {
     fi
 
     if [ "$platform" = "android" ]; then
-        cp docs/android.md INSTALL.md
+        cp docs/android_eclipse.md INSTALL_ECLIPSE.md
+        cp docs/android_studio.md INSTALL_ANDROID_STUDIO.md
     fi
     
     rm CONTRIBUTING.md
 
     #copy empty example
-    cd $pkg_ofroot/..
-    mkdir -p openFrameworks/apps/myApps 
+    cd $pkg_ofroot
+    mkdir -p apps/myApps 
     if [ "$pkg_platform" = "android" ]; then
-        cp -r openFrameworks/examples/android/androidEmptyExample openFrameworks/apps/myApps 
+        cp -r examples/android/androidEmptyExample apps/myApps/
     elif [ "$pkg_platform" = "ios" ]; then
-        cp -r openFrameworks/examples/ios/emptyExample openFrameworks/apps/myApps 
+        cp -r examples/ios/emptyExample apps/myApps/
     else
-        cp -r openFrameworks/examples/empty/emptyExample openFrameworks/apps/myApps 
+        cp -r examples/empty/emptyExample apps/myApps/
     fi
     
     #create compressed package
     if [ "$pkg_platform" = "linux" ] || [ "$pkg_platform" = "linux64" ] || [ "$pkg_platform" = "android" ] || [ "$pkg_platform" = "linuxarmv6l" ] || [ "$pkg_platform" = "linuxarmv7l" ]; then
+        echo "compressing package to of_v${pkg_version}_${pkg_platform}_release.tar.gz"
+        cd $pkg_ofroot/..
         mkdir of_v${pkg_version}_${pkg_platform}_release
-        mv openFrameworks/* of_v${pkg_version}_${pkg_platform}_release
+        mv ${pkgfolder}/* of_v${pkg_version}_${pkg_platform}_release
         COPYFILE_DISABLE=true tar czf of_v${pkg_version}_${pkg_platform}_release.tar.gz of_v${pkg_version}_${pkg_platform}_release
         rm -Rf of_v${pkg_version}_${pkg_platform}_release
     else
+        echo "compressing package to of_v${pkg_version}_${pkg_platform}_release.zip"
+        cd $pkg_ofroot/..
         mkdir of_v${pkg_version}_${pkg_platform}_release
-        mv openFrameworks/* of_v${pkg_version}_${pkg_platform}_release
+        mv ${pkgfolder}/* of_v${pkg_version}_${pkg_platform}_release
         zip -r of_v${pkg_version}_${pkg_platform}_release.zip of_v${pkg_version}_${pkg_platform}_release > /dev/null
-        mv of_v${pkg_version}_${pkg_platform}_release of_v${pkg_version}_${pkg_platform}_release        
         rm -Rf of_v${pkg_version}_${pkg_platform}_release
     fi
 }
 
+set -o pipefail  # trace ERR through pipes
+set -o errtrace  # trace ERR through 'time command' and other functions
+set -o nounset   # set -u : exit the script if you try to use an uninitialized variable
+set -o errexit   # set -e : exit the script if any statement returns a non-true return value
 
-if [ "$platform" = "all" ]; then
-    for eachplatform in win_cb linux linux64 vs osx 
-    do
-        cd $packageroot
-        mkdir of_v${version}_${eachplatform}
-        cp -R addons apps export libs other scripts of_v${version}_${eachplatform}
-        cd of_v${version}_${eachplatform}
-        createPackage $eachplatform $2 $PWD
-    done
-    
-    cd $packageroot
-    echo dir: $PWD
-    mkdir of_v${version}_all
-    mv addons apps export libs other scripts $packageroot/of_v${version}_all
-    COPYFILE_DISABLE=true tar czf of_v$version_all_FAT.tar.gz of_v${version}_all
-    rm -Rf of_v${version}_all
-    mv * $packageroot/..
-    #rm -Rf $packageroot
-else
-    of_root=$(cat ~/.ofprojectgenerator/config)
-    echo $packageroot > ~/.ofprojectgenerator/config
-    createPackage $platform $version $packageroot
-    echo $of_root > ~/.ofprojectgenerator/config
-    
-fi
+cleanup() {
+    cd $packageroot/..  
+    rm -rf ${pkgfolder} 
+}
+trap cleanup 0
 
+error() {
+  local parent_lineno="$1"
+  if [[ "$#" = "3" ]] ; then
+    local message="$2"
+    local code="${3:-1}"
+    echo "Error on or near line ${parent_lineno}: ${message}; exiting with status ${code}"
+  else
+    local code="${2:-1}"
+    echo "Error on or near line ${parent_lineno}; exiting with status ${code}"
+  fi
+  exit "${code}"
+}
+trap 'error ${LINENO}' ERR
 
-cd $packageroot/.. 
-rm -rf openFrameworks   
+createPackage $platform $version $packageroot $of_root    
+
+
